+++ NEWS.md
@@ -21,7 +21,7 @@
 ### Miscellaneous
 
 * Examples now use `scda.2022` rather than `scda.2021`.
-* Replaced deprecated `ggplot2` functions `..count..` and `..density..`.
+* Replaced deprecated `ggplot2` functions `..count..`, `..density..` and `..prop..`.
 
 # teal.modules.general 0.2.15